+++ kv/build.gradle
@@ -39,14 +39,11 @@ dependencies {
   testImplementation 'io.lettuce:lettuce-core'
   testImplementation 'org.fusesource.leveldbjni:leveldbjni-all'
   testImplementation  'org.infinispan:infinispan-core'
-  testImplementation 'org.jetbrains.spek:spek-api'
   testImplementation 'org.junit.jupiter:junit-jupiter-api'
   testImplementation 'org.junit.jupiter:junit-jupiter-params'
   testImplementation 'org.mapdb:mapdb'
   testImplementation 'org.rocksdb:rocksdbjni'
   testImplementation 'org.apache.openjpa:openjpa'
 
-
-  testRuntimeOnly 'org.jetbrains.spek:spek-junit-platform-engine'
   testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine'
 }
